Create index.php
This is just a skeleton of the homepage.  Most of the sections will likely wind up being developed as separate pages and included in the final version for potential reusability, but we'll see.

The meta and any other scripting is currently blank until we get a little further along and know exactly what we want to do, and how we want to present ourselves.