Create navigation.inc
This is the section where navigation will be generated. The code should add additional options based on the status of the user and whether or not they are signed in.

We don't need to (and shouldn't) include all of these things on the navigation right now, but I just wanted to have a braindump of all the possibilities.