Temps
To replace parts on the old site until the new one is functioning. The
links in index.php will need changing before going live. I just cut the
Roster and Schedule links. Plan to leave the about us and make it a
contact page as well. I'll send pictures for the slideshow at some
point. Fixed the background issue in the style sheet, isn't a radial
pattern but oh well.